+++ app/components/basic/ActiveButton/active_button.js
@@ -1,6 +1,6 @@
 import React from 'react';
 import PropTypes from 'prop-types';
-import '../basic_style.scss';
+import './style.scss';
 
 class ActiveButton extends React.PureComponent {
   render() {
