+++ app/components/basic/ActiveButton/style.scss
@@ -1,0 +1,11 @@
+@import '../../../common';
+
+.page-button_active {
+  background-color: $active_button_color;
+  border: none;
+  border-radius: 5px;
+  color: #ffffff;
+  font-weight: bold;
+  font-size: 1.6rem;
+  outline: none;
+}
