+++ app/components/basic/Button/style.scss
@@ -1,0 +1,24 @@
+@import '../../../common';
+
+.page-button {
+  background-color: $button_color;
+  border: none;
+  border-radius: 5px;
+  color: #ffffff;
+  font-weight: bold;
+  font-size: 1.6rem;
+  transition: 300ms;
+  outline: none;
+
+  &::first-letter {
+    text-transform: capitalize;
+  }
+
+  &:hover {
+      border: none;
+      border-radius: 5px;
+      background-color: #ffffff;
+      color: #530969;
+      font-weight: bold;
+  }
+}
