+++ app/components/basic/Icon/socialicon.js
@@ -1,6 +1,6 @@
 import React from 'react';
 import PropTypes from 'prop-types';
-import '../basic_style.scss';
+import './style.scss';
 
 class SocialIcon extends React.PureComponent {
   render() {
