+++ app/components/basic/Select/style.scss
@@ -1,0 +1,14 @@
+@import '../../../common';
+
+.select {
+  padding: 10px;
+  background-color: #ffffff;
+  border-radius: 5px;
+  -webkit-appearance: none;
+  outline: none;
+  font-size: 1.6rem;
+
+  option {
+    padding: 10px;
+  }
+}
