+++ app/components/basic/TextField/style.scss
@@ -1,0 +1,9 @@
+@import '../../../common';
+
+.field {
+  padding: 10px;
+  border-radius: 5px;
+  background-color: #ffffff;
+  font-size: 1.6rem;
+  outline: none;
+}
