+++ app/components/basic/ToggleButton/style.scss
@@ -1,0 +1,12 @@
+@import '../../../common';
+
+.toggle-button {
+  background-color: $button_color;
+  border: none;
+  border-radius: 5px;
+  color: #ffffff;
+  font-weight: bold;
+  text-transform: uppercase;
+  font-size: 1.6rem;
+  outline: none;
+}
