+++ app/components/basic/ToggleButton/toggleButton.js
@@ -1,6 +1,6 @@
 import React from 'react';
 import PropTypes from 'prop-types';
-import '../basic_style.scss';
+import './style.scss';
 
 class Toggle extends React.PureComponent {
   render() {
