--- app/components/basic/basic_style.scss
@@ -1,121 +1,0 @@
-@import '../../common';
-
-a {
-    text-decoration: none;
-}
-
-.page-button {
-  background-color: $button_color;
-  border: none;
-  border-radius: 5px;
-  color: #ffffff;
-  font-weight: bold;
-  font-size: 1.6rem;
-  transition: 300ms;
-  outline: none;
-
-  &::first-letter {
-    text-transform: capitalize;
-  }
-
-  &:hover {
-      border: none;
-      border-radius: 5px;
-      background-color: #ffffff;
-      color: #530969;
-      font-weight: bold;
-  }
-}
-
-.page-button_active {
-  background-color: $active_button_color;
-  border: none;
-  border-radius: 5px;
-  color: #ffffff;
-  font-weight: bold;
-  font-size: 1.6rem;
-  outline: none;
-}
-
-.toggle-button {
-  background-color: $button_color;
-  border: none;
-  border-radius: 5px;
-  color: #ffffff;
-  font-weight: bold;
-  text-transform: uppercase;
-  font-size: 1.6rem;
-  outline: none;
-}
-
-.field {
-  padding: 10px;
-  border-radius: 5px;
-  background-color: #ffffff;
-  font-size: 1.6rem;
-  outline: none;
-}
-
-.menu {
-  padding: 0px;
-  color: #ffffff;
-  font-weight: bolder;
-  font-size: 1.6rem;
-
-  li {
-    display: inline-block;
-    margin-right: 10px;
-    text-align: center;
-
-    &::after {
-      content: ' / ';
-      margin-left: 5px;
-    }
-
-    &:last-child::after {
-      content: '';
-    }
-
-    a {
-      outline: none;
-      color: #ffffff;
-    }
-
-  }
-
-}
-
-.socialicon {
-  display: block;
-  width: 43px;
-  height: 43px;
-  border-radius: 50%;
-  background-color: $button_color;
-  color: #ffffff;
-  font-size: 2.2rem;
-  text-align: center;
-  transition: 300ms;
-
-  &:hover {
-    width: 43px;
-    height: 43px;
-    border-radius: 50%;
-    background-color: #ffffff;
-    color:#530969;
-    font-size: 2.2rem;
-    text-align: center;
-  }
-}
-
-.select {
-  padding: 10px;
-  background-color: #ffffff;
-  border-radius: 5px;
-  -webkit-appearance: none;
-  outline: none;
-  font-size: 1.6rem;
-
-  option {
-    padding: 10px;
-  }
-}
